backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
pandas==2.2.3
scikit-learn==1.6.1
requests==2.32.3
numpy==2.2.4
gunicorn==23.0.0
python-dotenv==1.0.1
